move InsertValuesToken to rewrite package
